用户Store 和用户管理器

+++ src/ABPvNextOrangeAdmin.Domain/System/Roles/SysRoleMenu.cs
@@ -14,7 +14,7 @@ public class SysRoleMenu : FullAuditedEntity
     /// <summary>
     /// 菜单ID
     /// </summary>
-    public int MenuId { get; set; }
+    public long MenuId { get; set; }
 
     public override object[] GetKeys()
     {
